Finished Custom hooks and styles css
