+++ src/components/index.js
@@ -1,0 +1,4 @@
+//Archivo de Barril
+export * from './AddCategory';
+export * from './GifGrid';
+export * from './GifItem';
